1. 班级添加'老师'字段 2. fix bugs

+++ src/main/java/com/guanshan/phoenix/service/imp/TeacherServiceImp.java
@@ -195,9 +195,11 @@ public class TeacherServiceImp implements TeacherService {
         Teacher teacher = teacherMapper.selectByUserId(teacherId);
         if (teacher == null)
             throw new ApplicationErrorException(ErrorCode.TeacherNotExists);
-
         if (courseMapper.isTeacherUsedByCourse(teacherId)) {
             throw new ApplicationErrorException(ErrorCode.TeacherIsUsedByCourse);
+        }
+        if (clazzMapper.isTeacherUsedByClass(teacherId)) {
+            throw new ApplicationErrorException(ErrorCode.TeacherIsUsedByClass);
         }
 
         teacherMapper.deleteByUserId(teacherId);
